Relocate DomainController config page
