+++ root/usr/share/nethesis/NethServer/Template/SssdConfig/DomainController.php
